backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
gunicorn
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7
python-dotenv==1.0.1
sqlalchemy 
psycopg2
alembic
asyncpg
passlib
PyJWT
google-api-python-client
google-auth-httplib2
google-auth-oauthlib